Update Account posts to new response
