+++ frontend/next.config.ts
@@ -1,7 +1,15 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  images: {
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'konnectbuckettest.s3.amazonaws.com',
+        pathname: '**', // Allow any path
+      },
+    ],
+  }
 };
 
 export default nextConfig;
