+++ frontend/src/components/Feed/Feed.tsx
@@ -11,118 +11,118 @@ interface FeedProps {
 }
 
 // Mock data for Discover mode - all posts
-const discoverPosts: Post[] = [
-  {
-    id: '1',
-    author: { id: 'u1', username: 'alice', displayName: 'Alice Johnson', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Just finished my morning workout! 💪 Feeling energized and ready to tackle the day. What\'s everyone up to today?',
-    mediaType: 'text',
-    likesCount: 24,
-    isLiked: false,
-    createdAt: new Date(Date.now() - 1000 * 60 * 30).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 30).toISOString(),
-  },
-  {
-    id: '2',
-    author: { id: 'u2', username: 'bob', displayName: 'Bob Smith', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Amazing sunset at the beach today! 🌅',
-    mediaUrl: '/assets/images/placeholder.webp',
-    mediaType: 'image',
-    likesCount: 156,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
-  },
-  {
-    id: '3',
-    author: { id: 'u3', username: 'charlie', displayName: 'Charlie Brown', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Check out this cool trick! 🎯',
-    mediaUrl: '/sample.mp4',
-    mediaType: 'video',
-    likesCount: 89,
-    isLiked: false,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 4).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 4).toISOString(),
-  },
-  {
-    id: '4',
-    author: { id: 'u4', username: 'diana', displayName: 'Diana Prince', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Sometimes you just need to take a moment and appreciate the little things in life. Today I found this beautiful flower growing in the most unexpected place. It reminded me that beauty can be found anywhere if you look hard enough. 🌸',
-    mediaType: 'text',
-    likesCount: 42,
-    isLiked: false,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 6).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 6).toISOString(),
-  },
-  {
-    id: '5',
-    author: { id: 'u5', username: 'edward', displayName: 'Edward Norton', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Perfect timing! 📸',
-    mediaUrl: '/assets/images/placeholder.webp',
-    mediaType: 'image',
-    likesCount: 203,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
-  },
-  {
-    id: '6',
-    author: { id: 'u6', username: 'fiona', displayName: 'Fiona Gallagher', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'New recipe I tried today! 🍝',
-    mediaUrl: '/assets/images/placeholder.webp',
-    mediaType: 'image',
-    likesCount: 78,
-    isLiked: false,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 10).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 10).toISOString(),
-  },
-  {
-    id: '7',
-    author: { id: 'u7', username: 'george', displayName: 'George Washington', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Thought of the day: The best way to predict the future is to create it. What are you creating today? 🤔',
-    mediaType: 'text',
-    likesCount: 95,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
-  },
-];
+// const discoverPosts: Post[] = [
+//   {
+//     id: '1',
+//     author: { id: 'u1', username: 'alice', displayName: 'Alice Johnson', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Just finished my morning workout! 💪 Feeling energized and ready to tackle the day. What\'s everyone up to today?',
+//     mediaType: 'text',
+//     likesCount: 24,
+//     isLiked: false,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 30).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 30).toISOString(),
+//   },
+//   {
+//     id: '2',
+//     author: { id: 'u2', username: 'bob', displayName: 'Bob Smith', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Amazing sunset at the beach today! 🌅',
+//     mediaUrl: '/assets/images/placeholder.webp',
+//     mediaType: 'image',
+//     likesCount: 156,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
+//   },
+//   {
+//     id: '3',
+//     author: { id: 'u3', username: 'charlie', displayName: 'Charlie Brown', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Check out this cool trick! 🎯',
+//     mediaUrl: '/sample.mp4',
+//     mediaType: 'video',
+//     likesCount: 89,
+//     isLiked: false,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 4).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 4).toISOString(),
+//   },
+//   {
+//     id: '4',
+//     author: { id: 'u4', username: 'diana', displayName: 'Diana Prince', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Sometimes you just need to take a moment and appreciate the little things in life. Today I found this beautiful flower growing in the most unexpected place. It reminded me that beauty can be found anywhere if you look hard enough. 🌸',
+//     mediaType: 'text',
+//     likesCount: 42,
+//     isLiked: false,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 6).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 6).toISOString(),
+//   },
+//   {
+//     id: '5',
+//     author: { id: 'u5', username: 'edward', displayName: 'Edward Norton', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Perfect timing! 📸',
+//     mediaUrl: '/assets/images/placeholder.webp',
+//     mediaType: 'image',
+//     likesCount: 203,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
+//   },
+//   {
+//     id: '6',
+//     author: { id: 'u6', username: 'fiona', displayName: 'Fiona Gallagher', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'New recipe I tried today! 🍝',
+//     mediaUrl: '/assets/images/placeholder.webp',
+//     mediaType: 'image',
+//     likesCount: 78,
+//     isLiked: false,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 10).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 10).toISOString(),
+//   },
+//   {
+//     id: '7',
+//     author: { id: 'u7', username: 'george', displayName: 'George Washington', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Thought of the day: The best way to predict the future is to create it. What are you creating today? 🤔',
+//     mediaType: 'text',
+//     likesCount: 95,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
+//   },
+// ];
 
 // Mock data for Following mode - only posts from followed users
-const followingPosts: Post[] = [
-  {
-    id: '2',
-    author: { id: 'u2', username: 'bob', displayName: 'Bob Smith', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Amazing sunset at the beach today! 🌅',
-    mediaUrl: '/assets/images/placeholder.webp',
-    mediaType: 'image',
-    likesCount: 156,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
-  },
-  {
-    id: '5',
-    author: { id: 'u5', username: 'edward', displayName: 'Edward Norton', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Perfect timing! 📸',
-    mediaUrl: '/assets/images/placeholder.webp',
-    mediaType: 'image',
-    likesCount: 203,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
-  },
-  {
-    id: '7',
-    author: { id: 'u7', username: 'george', displayName: 'George Washington', avatarUrl: '/assets/images/missingAvatar.jpg' },
-    caption: 'Thought of the day: The best way to predict the future is to create it. What are you creating today? 🤔',
-    mediaType: 'text',
-    likesCount: 95,
-    isLiked: true,
-    createdAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
-    updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
-  },
-];
+// const followingPosts: Post[] = [
+//   {
+//     id: '2',
+//     author: { id: 'u2', username: 'bob', displayName: 'Bob Smith', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Amazing sunset at the beach today! 🌅',
+//     mediaUrl: '/assets/images/placeholder.webp',
+//     mediaType: 'image',
+//     likesCount: 156,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 2).toISOString(),
+//   },
+//   {
+//     id: '5',
+//     author: { id: 'u5', username: 'edward', displayName: 'Edward Norton', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Perfect timing! 📸',
+//     mediaUrl: '/assets/images/placeholder.webp',
+//     mediaType: 'image',
+//     likesCount: 203,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 8).toISOString(),
+//   },
+//   {
+//     id: '7',
+//     author: { id: 'u7', username: 'george', displayName: 'George Washington', avatarUrl: '/assets/images/missingAvatar.jpg' },
+//     caption: 'Thought of the day: The best way to predict the future is to create it. What are you creating today? 🤔',
+//     mediaType: 'text',
+//     likesCount: 95,
+//     isLiked: true,
+//     createdAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
+//     updatedAt: new Date(Date.now() - 1000 * 60 * 60 * 12).toISOString(),
+//   },
+// ];
 
 export function Feed({ mode }: FeedProps) {
   const [posts, setPosts] = useState<Post[]>([]);
@@ -240,7 +240,7 @@ export function Feed({ mode }: FeedProps) {
   }, [handleScroll]);
 
   const renderPost = (post: Post) => {
-    switch (post.mediaType) {
+    switch (post.media.type) {
       case 'text':
         return <TextPost key={post.id} post={post} />;
       case 'image':
